+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.16
-* added `url_to_absolute()` global function.
+* added `url_to_absolute()` global function;
+* removed `BodyParser::_turnUrlAsAbsolute()` method.
 
 ### 1.1.15
 * added `Entity` class.